ai_agents/data-collection/requirements.txt
# requirements.txt

python-dotenv
google-generativeai
fastapi
uvicorn
langgraph
bs4
tavily-python